refactor: Create containers for each major page section
My previous approach seems to be going down the path of being severely over-engineered. I don't know if that's geniuinely the case, or if I'm being insecure about the sheer number of times that I'm applying display: flex; to things.

This variant, in which I separate each section of the page, should help with that, though, we'll see.